Complete the code for RunAsync method

+++ MyExperiment/ExperimentResult.cs
@@ -41,7 +41,7 @@ namespace MyExperiment
         public string[] OutputFiles { get; set; }
         // Your properties related to experiment.
 
-        public float Accuracy { get; set; }
+        public double Accuracy { get; set; }
 
         public TimeSpan Duration { get; set; }
     }
